+++ CHANGES.md
@@ -5,6 +5,7 @@ CHANGES
 
 * add option parsing
 * add -i for querying cursor by specified interval in microseconds
+* add heatmap example
 
 ## Version 0.2 (2012-04-23T23:16:03Z)